Generator: Fix name conflict between declared class and generated classes.
If a declared class had a different namespace from, but the same
unqualified name as one of the generated query or model classes, the
generated use statement would create a name conflict between the declared
class and the base class, resulting in a PHP Fatal.

PHP Fatal error: Cannot use Name\DomainQuery as DomainQuery because
the name is already in use in /var/www/Name/Space/Base/Domain.php on
line <N>

This change makes sure we use an alias if the class name matches either the
base object or query class name.